Refactor MQTT into modular plugin projects
