--- DesktopApplicationTemplate.UI/Views/Mqtt/Advanced/MqttAdvancedConfigView.xaml.cs
@@ -1,22 +1,0 @@
-using System.Windows.Controls;
-using DesktopApplicationTemplate.Core.Services;
-using DesktopApplicationTemplate.UI.ViewModels.Mqtt.Advanced;
-
-namespace DesktopApplicationTemplate.UI.Views.Mqtt.Advanced;
-
-public partial class MqttAdvancedConfigView : Page
-{
-    private readonly ILoggingService _logger;
-
-    public MqttAdvancedConfigView(ILoggingService logger)
-    {
-        InitializeComponent();
-        _logger = logger;
-    }
-
-    public void Initialize(MqttAdvancedConfigViewModel vm)
-    {
-        DataContext = vm;
-        vm.Logger = _logger;
-    }
-}
