--- DesktopApplicationTemplate.UI/Views/Mqtt/Create/MqttCreateServiceView.xaml
@@ -1,102 +1,0 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Mqtt.Create.MqttCreateServiceView"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:svc="clr-namespace:DesktopApplicationTemplate.UI.Services"
-      xmlns:views="clr-namespace:DesktopApplicationTemplate.UI.Views"
-      mc:Ignorable="d">
-    <ScrollViewer VerticalScrollBarVisibility="Auto">
-        <Grid Margin="20">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" /> <!-- Service Name -->
-                <RowDefinition Height="Auto" /> <!-- Host -->
-                <RowDefinition Height="Auto" /> <!-- Port -->
-                <RowDefinition Height="Auto" /> <!-- ClientId -->
-                <RowDefinition Height="Auto" /> <!-- Username -->
-                <RowDefinition Height="Auto" /> <!-- Password -->
-                <RowDefinition Height="Auto" /> <!-- Connection Type -->
-                <RowDefinition Height="Auto" /> <!-- Buttons -->
-            </Grid.RowDefinitions>
-
-            <TextBlock Grid.Row="0" Grid.Column="0" Text="Service Name" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="0" Grid.Column="1" Text="{Binding ServiceName, ValidatesOnNotifyDataErrors=True, UpdateSourceTrigger=PropertyChanged}">
-                <TextBox.Style>
-                    <Style TargetType="TextBox" BasedOn="{StaticResource FormField}">
-                        <Style.Triggers>
-                            <Trigger Property="Validation.HasError" Value="True">
-                                <Setter Property="ToolTip" Value="{Binding (Validation.Errors)[0].ErrorContent, RelativeSource={RelativeSource Self}}"/>
-                            </Trigger>
-                        </Style.Triggers>
-                    </Style>
-                </TextBox.Style>
-            </TextBox>
-
-            <TextBlock Grid.Row="1" Grid.Column="0" Text="Host" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="1" Grid.Column="1" Text="{Binding Host, ValidatesOnNotifyDataErrors=True, UpdateSourceTrigger=PropertyChanged}">
-                <TextBox.Style>
-                    <Style TargetType="TextBox" BasedOn="{StaticResource FormField}">
-                        <Style.Triggers>
-                            <Trigger Property="Validation.HasError" Value="True">
-                                <Setter Property="ToolTip" Value="{Binding (Validation.Errors)[0].ErrorContent, RelativeSource={RelativeSource Self}}"/>
-                            </Trigger>
-                        </Style.Triggers>
-                    </Style>
-                </TextBox.Style>
-            </TextBox>
-
-            <TextBlock Grid.Row="2" Grid.Column="0" Text="Port" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="2" Grid.Column="1" Text="{Binding Port, ValidatesOnNotifyDataErrors=True, UpdateSourceTrigger=PropertyChanged}">
-                <TextBox.Style>
-                    <Style TargetType="TextBox" BasedOn="{StaticResource FormField}">
-                        <Style.Triggers>
-                            <Trigger Property="Validation.HasError" Value="True">
-                                <Setter Property="ToolTip" Value="{Binding (Validation.Errors)[0].ErrorContent, RelativeSource={RelativeSource Self}}"/>
-                            </Trigger>
-                        </Style.Triggers>
-                    </Style>
-                </TextBox.Style>
-            </TextBox>
-
-            <TextBlock Grid.Row="3" Grid.Column="0" Text="Client Id" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="3" Grid.Column="1" Text="{Binding ClientId, ValidatesOnNotifyDataErrors=True, UpdateSourceTrigger=PropertyChanged}">
-                <TextBox.Style>
-                    <Style TargetType="TextBox" BasedOn="{StaticResource FormField}">
-                        <Style.Triggers>
-                            <Trigger Property="Validation.HasError" Value="True">
-                                <Setter Property="ToolTip" Value="{Binding (Validation.Errors)[0].ErrorContent, RelativeSource={RelativeSource Self}}"/>
-                            </Trigger>
-                        </Style.Triggers>
-                    </Style>
-                </TextBox.Style>
-            </TextBox>
-
-            <TextBlock Grid.Row="4" Grid.Column="0" Text="Username" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="4" Grid.Column="1" Text="{Binding Username}" Style="{StaticResource FormField}" ToolTip="Username"/>
-
-            <TextBlock Grid.Row="5" Grid.Column="0" Text="Password" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="5" Grid.Column="1" Text="{Binding Password}" Style="{StaticResource FormField}" ToolTip="Password"/>
-
-            <TextBlock Grid.Row="6" Grid.Column="0" Text="Connection Type" Style="{StaticResource FormLabel}"/>
-            <ComboBox Grid.Row="6" Grid.Column="1" SelectedValue="{Binding Options.ConnectionType}" SelectedValuePath="Tag" Style="{StaticResource FormField}">
-                <ComboBoxItem Content="MQTT" Tag="{x:Static svc:MqttConnectionType.Mqtt}"/>
-                <ComboBoxItem Content="MQTT TLS" Tag="{x:Static svc:MqttConnectionType.MqttTls}"/>
-                <ComboBoxItem Content="WebSocket" Tag="{x:Static svc:MqttConnectionType.WebSocket}"/>
-                <ComboBoxItem Content="WebSocket TLS" Tag="{x:Static svc:MqttConnectionType.WebSocketTls}"/>
-            </ComboBox>
-
-            <views:EditorButtonBar Grid.Row="7" Grid.ColumnSpan="2"
-                                   AdvancedConfigCommand="{Binding AdvancedConfigCommand}"
-                                   SaveCommand="{Binding CreateCommand}"
-                                   CancelCommand="{Binding CancelCommand}"
-                                   SaveButtonText="Create"
-                                   AdvancedAutomationName="Open MQTT Advanced Configuration"
-                                   SaveAutomationName="Create MQTT Service"
-                                   CancelAutomationName="Cancel MQTT Creation" />
-        </Grid>
-    </ScrollViewer>
-</Page>
