--- DesktopApplicationTemplate.UI/Views/Mqtt/Create/MqttCreateServiceView.xaml.cs
@@ -1,15 +1,0 @@
-using System.Windows.Controls;
-using DesktopApplicationTemplate.Core.Services;
-using DesktopApplicationTemplate.UI.ViewModels.Mqtt.Create;
-
-namespace DesktopApplicationTemplate.UI.Views.Mqtt.Create;
-
-public partial class MqttCreateServiceView : Page
-{
-    public MqttCreateServiceView(MqttCreateServiceViewModel vm, ILoggingService logger)
-    {
-        InitializeComponent();
-        DataContext = vm;
-        vm.Logger = logger;
-    }
-}
