--- DesktopApplicationTemplate.UI/Views/Mqtt/Edit/MqttEditConnectionView.xaml
@@ -1,97 +1,0 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Mqtt.Edit.MqttEditConnectionView"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:svc="clr-namespace:DesktopApplicationTemplate.UI.Services"
-      mc:Ignorable="d"
-      xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006">
-    <ScrollViewer VerticalScrollBarVisibility="Auto">
-        <Grid Margin="20">
-            <Grid.ColumnDefinitions>
-                <ColumnDefinition Width="Auto" />
-                <ColumnDefinition Width="*" />
-            </Grid.ColumnDefinitions>
-            <Grid.RowDefinitions>
-                <RowDefinition Height="Auto" /> <!-- Host -->
-                <RowDefinition Height="Auto" /> <!-- Port -->
-                <RowDefinition Height="Auto" /> <!-- ClientId -->
-                <RowDefinition Height="Auto" /> <!-- Username -->
-                <RowDefinition Height="Auto" /> <!-- Password -->
-                <RowDefinition Height="Auto" /> <!-- Connection Type -->
-                <RowDefinition Height="Auto" /> <!-- WebSocket Path -->
-                <RowDefinition Height="Auto" /> <!-- Buttons -->
-            </Grid.RowDefinitions>
-
-            <TextBlock Grid.Row="0" Grid.Column="0" Text="Host" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="0" Grid.Column="1" Text="{Binding Host}" Style="{StaticResource FormField}"/>
-
-            <TextBlock Grid.Row="1" Grid.Column="0" Text="Port" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="1" Grid.Column="1" Text="{Binding Port}" Style="{StaticResource FormField}"/>
-
-            <TextBlock Grid.Row="2" Grid.Column="0" Text="Client Id" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="2" Grid.Column="1" Text="{Binding ClientId}" Style="{StaticResource FormField}"/>
-
-            <TextBlock Grid.Row="3" Grid.Column="0" Text="Username" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="3" Grid.Column="1" Text="{Binding Username}" Style="{StaticResource FormField}"/>
-
-            <TextBlock Grid.Row="4" Grid.Column="0" Text="Password" Style="{StaticResource FormLabel}"/>
-            <TextBox Grid.Row="4" Grid.Column="1" Text="{Binding Password}" Style="{StaticResource FormField}"/>
-
-            <TextBlock Grid.Row="5" Grid.Column="0" Text="Connection Type" Style="{StaticResource FormLabel}"/>
-            <ComboBox Grid.Row="5" Grid.Column="1" SelectedValue="{Binding ConnectionType}" SelectedValuePath="Tag" Style="{StaticResource FormField}">
-                <ComboBoxItem Content="MQTT" Tag="{x:Static svc:MqttConnectionType.Mqtt}"/>
-                <ComboBoxItem Content="MQTT TLS" Tag="{x:Static svc:MqttConnectionType.MqttTls}"/>
-                <ComboBoxItem Content="WebSocket" Tag="{x:Static svc:MqttConnectionType.WebSocket}"/>
-                <ComboBoxItem Content="WebSocket TLS" Tag="{x:Static svc:MqttConnectionType.WebSocketTls}"/>
-            </ComboBox>
-
-            <TextBlock Grid.Row="6" Grid.Column="0" Text="WS Path">
-                <TextBlock.Style>
-                    <Style TargetType="TextBlock" BasedOn="{StaticResource FormLabel}">
-                        <Setter Property="Visibility" Value="Collapsed" />
-                        <Style.Triggers>
-                            <DataTrigger Binding="{Binding ConnectionType}" Value="WebSocket">
-                                <Setter Property="Visibility" Value="Visible" />
-                            </DataTrigger>
-                            <DataTrigger Binding="{Binding ConnectionType}" Value="WebSocketTls">
-                                <Setter Property="Visibility" Value="Visible" />
-                            </DataTrigger>
-                        </Style.Triggers>
-                    </Style>
-                </TextBlock.Style>
-            </TextBlock>
-            <TextBox Grid.Row="6" Grid.Column="1" Text="{Binding WebSocketPath}">
-                <TextBox.Style>
-                    <Style TargetType="TextBox" BasedOn="{StaticResource FormField}">
-                        <Setter Property="Visibility" Value="Collapsed" />
-                        <Style.Triggers>
-                            <DataTrigger Binding="{Binding ConnectionType}" Value="WebSocket">
-                                <Setter Property="Visibility" Value="Visible" />
-                            </DataTrigger>
-                            <DataTrigger Binding="{Binding ConnectionType}" Value="WebSocketTls">
-                                <Setter Property="Visibility" Value="Visible" />
-                            </DataTrigger>
-                        </Style.Triggers>
-                    </Style>
-                </TextBox.Style>
-            </TextBox>
-            
-            <StackPanel Grid.Row="7" Grid.ColumnSpan="2" Orientation="Horizontal" HorizontalAlignment="Center" Margin="0,10,0,0">
-                <Button x:Name="SubscriptionButton" Content="{Binding SubscriptionButtonText}" Width="110" Margin="5" Command="{Binding ToggleSubscriptionCommand}" AutomationProperties.Name="{Binding SubscriptionButtonText}">
-                    <Button.Style>
-                        <Style TargetType="Button">
-                            <Setter Property="Background" Value="Green"/>
-                            <Style.Triggers>
-                                <DataTrigger Binding="{Binding IsConnected}" Value="True">
-                                    <Setter Property="Background" Value="Red"/>
-                                </DataTrigger>
-                            </Style.Triggers>
-                        </Style>
-                    </Button.Style>
-                </Button>
-                <Button x:Name="CancelButton" Content="Cancel" Width="90" Margin="5" Command="{Binding CancelCommand}" AutomationProperties.Name="Cancel Update"/>
-                <Button x:Name="UpdateButton" Content="Update" Width="90" Margin="5" Command="{Binding UpdateCommand}" AutomationProperties.Name="Update Connection"/>
-            </StackPanel>
-        </Grid>
-    </ScrollViewer>
-</Page>
