--- DesktopApplicationTemplate.UI/Views/Mqtt/Edit/MqttEditConnectionView.xaml.cs
@@ -1,13 +1,0 @@
-using System.Windows.Controls;
-using DesktopApplicationTemplate.UI.ViewModels.Mqtt.Edit;
-
-namespace DesktopApplicationTemplate.UI.Views.Mqtt.Edit;
-
-public partial class MqttEditConnectionView : Page
-{
-    public MqttEditConnectionView(MqttEditConnectionViewModel vm)
-    {
-        InitializeComponent();
-        DataContext = vm;
-    }
-}
