--- DesktopApplicationTemplate.UI/Views/Mqtt/MqttTagSubscriptionsView.xaml
@@ -1,116 +1,0 @@
-<Page x:Class="DesktopApplicationTemplate.UI.Views.Mqtt.MqttTagSubscriptionsView"
-      xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-      xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-      xmlns:models="clr-namespace:DesktopApplicationTemplate.UI.Models"
-      xmlns:mqtt="clr-namespace:MQTTnet.Protocol;assembly=MQTTnet"
-      xmlns:d="http://schemas.microsoft.com/expression.blend/2008"
-      xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-      xmlns:sys="clr-namespace:System;assembly=System.Runtime"
-      xmlns:behaviors="clr-namespace:DesktopApplicationTemplate.UI.Behaviors"
-      xmlns:helpers="clr-namespace:DesktopApplicationTemplate.UI.Helpers"
-      xmlns:shared="clr-namespace:DesktopApplicationTemplate.UI.Views.Shared"
-      mc:Ignorable="d">
-    <Page.Resources>
-        <ObjectDataProvider x:Key="QoSValues" MethodName="GetValues" ObjectType="{x:Type sys:Enum}">
-            <ObjectDataProvider.MethodParameters>
-                <x:Type TypeName="mqtt:MqttQualityOfServiceLevel" />
-            </ObjectDataProvider.MethodParameters>
-        </ObjectDataProvider>
-        <helpers:BooleanToBrushConverter x:Key="BoolToBrush" TrueBrush="Green" FalseBrush="Red" />
-        <helpers:StringNullOrEmptyToVisibilityConverter x:Key="StringNullOrEmptyToVisibilityConverter" />
-        <helpers:NullToVisibilityConverter x:Key="NullToVisibilityConverter" />
-    </Page.Resources>
-    <Grid Margin="20">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-        </Grid.RowDefinitions>
-        <Border Grid.Row="0" CornerRadius="8" BorderThickness="1" Padding="10" Margin="0,0,0,10"
-                Background="{Binding Tag.BackgroundColor, RelativeSource={RelativeSource AncestorType=Page}}"
-                BorderBrush="{Binding Tag.BorderColor, RelativeSource={RelativeSource AncestorType=Page}}"
-                ToolTip="{Binding Tag.DescriptorTooltip, RelativeSource={RelativeSource AncestorType=Page}}">
-            <DockPanel>
-                <Border Width="40" Height="40" CornerRadius="20" Margin="0,0,10,0"
-                        Background="{Binding Tag.BorderColor, RelativeSource={RelativeSource AncestorType=Page}}"
-                        Visibility="{Binding Tag.IconGlyph, RelativeSource={RelativeSource AncestorType=Page}, Converter={StaticResource NullToVisibilityConverter}}">
-                    <TextBlock Text="{Binding Tag.IconGlyph, RelativeSource={RelativeSource AncestorType=Page}}"
-                               FontSize="22" HorizontalAlignment="Center" VerticalAlignment="Center" />
-                </Border>
-                <StackPanel>
-                    <TextBlock Text="{Binding Tag.DisplayName, RelativeSource={RelativeSource AncestorType=Page}}" FontWeight="Bold" />
-                    <TextBlock Text="{Binding Tag.DescriptorDescription, RelativeSource={RelativeSource AncestorType=Page}}"
-                               Foreground="Gray" TextWrapping="Wrap"
-                               Visibility="{Binding Tag.DescriptorDescription, RelativeSource={RelativeSource AncestorType=Page}, Converter={StaticResource StringNullOrEmptyToVisibilityConverter}}" />
-                </StackPanel>
-            </DockPanel>
-        </Border>
-        <!-- Connection bar -->
-        <StackPanel Grid.Row="1" Orientation="Horizontal" Margin="0,0,0,10">
-            <Button Content="Connect" Command="{Binding ConnectCommand}" Width="80" AutomationProperties.Name="Connect MQTT"/>
-            <Ellipse Width="10" Height="10" Margin="5,0,0,0" Fill="{Binding IsConnected, Converter={StaticResource BoolToBrush}}"/>
-            <TextBox Text="{Binding NewTopic}" Width="200" Margin="10,0,0,0" x:Name="NewTopicBox" ToolTip="Topic to subscribe"
-                     behaviors:TextBoxHintBehavior.AutoToolTip="True"/>
-            <ComboBox Width="100" Margin="5,0,0,0"
-                      ItemsSource="{Binding Source={StaticResource QoSValues}}"
-                      SelectedItem="{Binding NewQoS}"/>
-            <Button Content="Add"
-                    Command="{Binding AddTopicCommand}"
-                    Width="50"
-                    Margin="5,0,0,0"
-                    IsEnabled="{Binding CanAddTopic}"/>
-            <Button Content="Remove" Command="{Binding RemoveTopicCommand}" Width="70" Margin="5,0,0,0"/>
-        </StackPanel>
-        <!-- Subscriptions list -->
-        <DataGrid Grid.Row="2"
-                  ItemsSource="{Binding Subscriptions}"
-                  SelectedItem="{Binding SelectedSubscription}"
-                  AutoGenerateColumns="False"
-                  HeadersVisibility="Column"
-                  CanUserAddRows="False">
-            <DataGrid.Columns>
-                <DataGridTextColumn Header="Topic" Binding="{Binding Topic}" IsReadOnly="True"/>
-                <DataGridComboBoxColumn Header="QoS"
-                                        SelectedItemBinding="{Binding QoS}"
-                                        ItemsSource="{Binding Source={StaticResource QoSValues}}"/>
-                <DataGridTextColumn Header="Endpoint"
-                                    Binding="{Binding Endpoint, UpdateSourceTrigger=PropertyChanged}"/>
-                <DataGridTextColumn Header="Message"
-                                    Binding="{Binding OutgoingMessage, UpdateSourceTrigger=PropertyChanged}"/>
-                <DataGridTemplateColumn Header="Test">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate DataType="{x:Type models:TagSubscription}">
-                            <Button Content="Test" Width="50"
-                                    Command="{Binding DataContext.TestTagEndpointCommand, RelativeSource={RelativeSource AncestorType=DataGrid}}"
-                                    CommandParameter="{Binding}"/>
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
-            </DataGrid.Columns>
-        </DataGrid>
-        <!-- Test message bar -->
-        <StackPanel Grid.Row="3" Orientation="Horizontal" Margin="0,10,0,0">
-            <TextBox Text="{Binding SelectedSubscription.OutgoingMessage, UpdateSourceTrigger=PropertyChanged}"
-                     Width="200"
-                     x:Name="TestMessageBox"
-                     ToolTip="Message to send"
-                     behaviors:TextBoxHintBehavior.AutoToolTip="True"/>
-            <Button x:Name="SendButton"
-                    Content="Send"
-                    Command="{Binding PublishTestMessageCommand}"
-                    Width="60"
-                    Margin="5,0,0,0"
-                    AutomationProperties.Name="Send Test Message"/>
-            <Button Content="Test"
-                    Command="{Binding TestTagEndpointCommand}"
-                    CommandParameter="{Binding SelectedSubscription}"
-                    Width="60"
-                    Margin="5,0,0,0"/>
-        </StackPanel>
-
-        <!-- Log entries -->
-        <shared:ServiceLogView Grid.Row="4" Margin="0,10,0,0" x:Name="LogView" />
-    </Grid>
-</Page>
