+++ DesktopApplicationTemplate.sln
@@ -23,6 +23,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "HttpRelayPlugin", "Samples\
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ServicePluginTemplate", "Templates\ServicePluginTemplate\ServicePluginTemplate.csproj", "{1ED21731-2F3F-4A42-A9BD-8EA34C705B1C}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MqttPackage", "Services\Mqtt\MqttPackage\MqttPackage.csproj", "{533D7FA1-8C3E-49AC-ACA2-E0D95F2C764D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "MqttUI", "Services\Mqtt\MqttUI\MqttUI.csproj", "{9D22FB33-5081-418F-961A-20CFEFD1C584}"
+EndProject
 
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -70,6 +74,14 @@ Global
                 {1ED21731-2F3F-4A42-A9BD-8EA34C705B1C}.Debug|Any CPU.Build.0 = Debug|Any CPU
                 {1ED21731-2F3F-4A42-A9BD-8EA34C705B1C}.Release|Any CPU.ActiveCfg = Release|Any CPU
                 {1ED21731-2F3F-4A42-A9BD-8EA34C705B1C}.Release|Any CPU.Build.0 = Release|Any CPU
+                {533D7FA1-8C3E-49AC-ACA2-E0D95F2C764D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {533D7FA1-8C3E-49AC-ACA2-E0D95F2C764D}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {533D7FA1-8C3E-49AC-ACA2-E0D95F2C764D}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {533D7FA1-8C3E-49AC-ACA2-E0D95F2C764D}.Release|Any CPU.Build.0 = Release|Any CPU
+                {9D22FB33-5081-418F-961A-20CFEFD1C584}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+                {9D22FB33-5081-418F-961A-20CFEFD1C584}.Debug|Any CPU.Build.0 = Debug|Any CPU
+                {9D22FB33-5081-418F-961A-20CFEFD1C584}.Release|Any CPU.ActiveCfg = Release|Any CPU
+                {9D22FB33-5081-418F-961A-20CFEFD1C584}.Release|Any CPU.Build.0 = Release|Any CPU
         EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
