+++ Services/Mqtt/MqttUI/MqttUI.csproj
@@ -1,0 +1,37 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <Import Project="..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.props" Condition="Exists('..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.props')" />
+  <PropertyGroup>
+    <TargetFramework>net8.0-windows</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <UseWPF>true</UseWPF>
+    <RootNamespace>DesktopApplicationTemplate.Services.Mqtt.UI</RootNamespace>
+    <AssemblyName>DesktopApplicationTemplate.Mqtt.UI</AssemblyName>
+    <ServicePluginIncludeReferenceCopyLocal>false</ServicePluginIncludeReferenceCopyLocal>
+    <ServicePluginPackageId>DesktopApplicationTemplate.Mqtt.UI</ServicePluginPackageId>
+    <ServicePluginPackageVersion>1.0.0</ServicePluginPackageVersion>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\..\DesktopApplicationTemplate.Core\DesktopApplicationTemplate.Core.csproj" />
+    <ProjectReference Include="..\..\MqttPackage\MqttPackage.csproj" />
+    <ProjectReference Include="..\..\..\DesktopApplicationTemplate.UI\DesktopApplicationTemplate.UI.csproj" />
+    <ProjectReference Include="..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.csproj" ReferenceOutputAssembly="false" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="MQTTnet" Version="4.3.7.1207" />
+    <PackageReference Include="Microsoft.Extensions.Options.ConfigurationExtensions" Version="8.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <Page Update="Views\**\*.xaml">
+      <Generator>MSBuild:Compile</Generator>
+      <SubType>Designer</SubType>
+    </Page>
+    <Compile Update="Views\**\*.xaml.cs">
+      <DependentUpon>%(Filename).xaml</DependentUpon>
+    </Compile>
+  </ItemGroup>
+  <Import Project="..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.targets" Condition="Exists('..\..\..\ServicePlugin.Packaging\ServicePlugin.Packaging.targets')" />
+</Project>
