+++ Services/Mqtt/MqttUI/Services/MqttServiceOptions.cs
@@ -1,0 +1,32 @@
+using System;
+using MQTTnet.Protocol;
+
+namespace DesktopApplicationTemplate.Services.Mqtt.UI.Services
+{
+    public enum MqttConnectionType
+    {
+        Mqtt,
+        MqttTls,
+        WebSocket,
+        WebSocketTls
+    }
+
+    public class MqttServiceOptions
+    {
+        public string Host { get; set; } = string.Empty;
+        public int Port { get; set; } = 1883;
+        public string ClientId { get; set; } = string.Empty;
+        public string? Username { get; set; }
+        public string? Password { get; set; }
+        public MqttConnectionType ConnectionType { get; set; } = MqttConnectionType.Mqtt;
+        public string? WebSocketPath { get; set; }
+        public byte[]? ClientCertificate { get; set; }
+        public string? WillTopic { get; set; }
+        public string? WillPayload { get; set; }
+        public MqttQualityOfServiceLevel WillQualityOfService { get; set; } = MqttQualityOfServiceLevel.AtMostOnce;
+        public bool WillRetain { get; set; }
+        public ushort KeepAliveSeconds { get; set; } = 60;
+        public bool CleanSession { get; set; } = true;
+        public TimeSpan? ReconnectDelay { get; set; }
+    }
+}
