+++ Services/Mqtt/MqttUI/Views/Mqtt/MqttTagSubscriptionsView.xaml.cs
@@ -1,0 +1,33 @@
+using System;
+using System.Windows.Controls;
+using DesktopApplicationTemplate.Core.Services;
+using DesktopApplicationTemplate.Models;
+using DesktopApplicationTemplate.Services.Mqtt.Descriptors;
+using DesktopApplicationTemplate.Services.Mqtt.UI.ViewModels.Mqtt;
+using DesktopApplicationTemplate.UI.ViewModels;
+using DesktopApplicationTemplate.UI.Configuration;
+
+namespace DesktopApplicationTemplate.Services.Mqtt.UI.Views.Mqtt;
+
+/// <summary>
+/// Interaction logic for MqttTagSubscriptionsView.xaml
+/// </summary>
+[ServiceDescriptorRegistration(MqttServiceDescriptor.DescriptorId, ServiceRegistrationKind.ServicePage)]
+public partial class MqttTagSubscriptionsView : Page, IServiceLogHost
+{
+    /// <summary>
+    /// Initializes a new instance of the <see cref="MqttTagSubscriptionsView"/> class.
+    /// </summary>
+    public MqttTagSubscriptionsView(MqttTagSubscriptionsViewModel vm, ILoggingService logger)
+    {
+        InitializeComponent();
+        vm.Logger = logger;
+        DataContext = vm;
+    }
+
+    /// <inheritdoc />
+    public void SetServiceContext(ServiceListModel service)
+    {
+        LogView.DataContext = new ServiceLogViewModel(service.Type, service.Logs);
+    }
+}
