+++ Services/Mqtt/MqttUI/plugin.manifest.json
@@ -1,0 +1,12 @@
+{
+  "id": "DesktopApplicationTemplate.Mqtt.UI",
+  "name": "MQTT Service UI",
+  "version": "1.0.0",
+  "entryAssembly": "DesktopApplicationTemplate.Mqtt.UI.dll",
+  "serviceAssemblies": [
+    "DesktopApplicationTemplate.Mqtt.UI.dll"
+  ],
+  "probingPaths": [
+    "libs"
+  ]
+}
